moved the data classes to separate files independent of the UI
