+++ lib/src/features/home_store/domain/category_data.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class Category {
+  final IconData icon;
+  final String title;
+
+  const Category({required this.icon, required this.title});
+}
