+++ lib/src/features/home_store/domain/phone_item.dart
@@ -1,0 +1,13 @@
+class PhoneItem {
+  final String image;
+  final String currentPrice;
+  final String previousPrice;
+  final String name;
+
+  PhoneItem({
+    required this.image,
+    required this.currentPrice,
+    required this.previousPrice,
+    required this.name,
+  });
+}
