+++ lib/src/features/home_store/presentation/ui/widget/category_text_widget.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
+import '../../../domain/category_data.dart';
 
 
 class CategoryData {
-  late final bool isSelected;
+
+  final bool isSelected;
   final Icon icon;
 
   CategoryData({required this.isSelected, required this.icon});
@@ -29,11 +31,11 @@ class CategoryTextWidgetState extends State<CategoryTextWidget> {
     });
   }
 
-  List<Category> categories = [
-    Category(icon: Icons.smartphone_outlined, title: 'Phones'),
-    Category(icon: Icons.computer_outlined, title: 'Computer'),
-    Category(icon: Icons.health_and_safety_outlined, title: 'Health'),
-    Category(icon: Icons.book_outlined, title: 'Books'),
+  final List<Category> categories = [
+    const Category(icon: Icons.smartphone_outlined, title: 'Phones'),
+    const Category(icon: Icons.computer_outlined, title: 'Computer'),
+    const Category(icon: Icons.health_and_safety_outlined, title: 'Health'),
+    const Category(icon: Icons.book_outlined, title: 'Books'),
   ];
 
   @override
@@ -81,7 +83,7 @@ class _CategoryItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final color = isSelected ? const Color(0xFFFF6E4E) : Colors.white;
+    final color = isSelected ? const Color(0xFF010035) : Colors.white;
 
     return Column(
       children: [
@@ -117,11 +119,4 @@ class _CategoryItem extends StatelessWidget {
       ],
     );
   }
-}
-
-class Category {
-  final IconData icon;
-  final String title;
-
-  Category({required this.icon, required this.title});
 }
